Fix html5 error clearing test failures due to incomplete mocks, migrate HTML5 error handling into HTML5 provider
